requirements.txt
django==3.2.7
django-ckeditor==6.1.0
Pillow==8.3.2
pylint==2.10.2
pylint-celery==0.3
pylint-django==2.4.4
python-dotenv==0.19.0